requirements.txt
APScheduler
arabic-reshaper
asgiref
asn1crypto
certifi
cffi
chardet
charset-normalizer
click
cryptography
cssselect2
Django>=4.2.11,<5.0
django-apscheduler
django-cors-headers
django-environ
django-filter
django-haystack
django-jsonfield
django-mathfilters
django-model-utils
django-simple-history
django-widget-tweaks
et-xmlfile
html5lib
idna
lxml
numpy
openpyxl
oscrypto
pandas
pillow
pycparser
pyHanko
pyhanko-certvalidator
pypdf
pypng
python-bidi
python-dateutil
pytz
PyYAML
Pyzk
qrcode
reportlab
requests
setuptools
six
sqlparse
svglib
swapper
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
utils
webencodings
Whoosh
xhtml2pdf
XlsxWriter
gunicorn
psycopg2-binary
whitenoise
